+++ src/components/auth/AuthForm.jsx
@@ -3,9 +3,10 @@ import { AUTH_MODE, BUTTON_MODE } from '../../constants/mode';
 import { Button } from '../common/Button';
 import { InputBar, InputRadio } from '../common/Input';
 import { Link } from 'react-router-dom';
-import { PATH } from '../../constants/RouterPathConstants';
+
 import ErrorText from '../common/ErrorText';
 import SignupAddressSelect from './SignupAddressInput';
+import { PATH } from '../../constants/routerPath';
 
 const AuthForm = ({
   mode,
